ADD SCROLL && MASSIVE BUFFER AND SCREEN UPDATE:
The buffer is now used only in order to change screens.
Otherwise everything is done straight on the vga.

The iteration is now done on an array of shorts using masks and
bitshift operations for chars and colours.

The scroll is working and implemented.
Checks are done on the arrow keys in order to control where the cursor
may or may not end up. Same for deletion. We are now only able to delete
the "current prompt".

The default pattern for new characters midline is insert.
A toggle could be added with key insert in order to decide on which
behaviour to adopt.

Newline added mid-line is not dealt with as insert yet and will erase
the rest of the line.

Still need to work on lidt implementation to have proper interrupt
management.